add Capsule
- move positive_mod into geom::
- add some hash util functions
- add hashes for everybody
- resize functions for AffineXf
- more rotation matrix utility fns
- improve ctors for SimpleMatrix (more initializer_lists, e.g.)
- get_angle() -> angle()
- more ctors for Vec + syntax cleanup
- add Capsule
- add roundrect() convenience function
- ctors -> explicit in Oriented
- OrientedRect -> fwd decl
- initializer_list for Storage classes
- make install sudo-safe

+++ geomc/linalg/Quaternion.h
@@ -328,7 +328,7 @@ namespace std {
 template <typename T>
 struct hash<geom::Quat<T>> {
     size_t operator()(const geom::Quat<T> &v) const {
-        constexpr size_t nonce = 0x25ca848a1305d3a2ULL;
+        constexpr size_t nonce = (size_t) 0x25ca848a1305d3a2ULL;
         return geom::hash_bytes(&v, sizeof(geom::Quat<T>)) ^ nonce;
     }
 };
